dao test integracion corregida

+++ src/main/java/com/distribuida/dao/FacturaDetalleRepository.java
@@ -3,6 +3,6 @@ package com.distribuida.dao;
 import com.distribuida.model.FacturaDetalle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacturaDetalleRepository extends JpaRepository<FacturaDetalle, Long> {
+public interface FacturaDetalleRepository extends JpaRepository<FacturaDetalle, Integer> {
 
 }
